Moving IRC and Twitter into integrations folder
