--- lib/integrations/integrations.js
@@ -1,11 +1,0 @@
-(function() {
-    var define = require('amdefine')(module);
-
-    var deps = [];
-
-    define(deps, function() {
-       module.exports = {
-
-       };
-    });
-}());
